Add self-hosted package publishing (#5)
## Motivation for the change, related issues

We need to deploy JS packages to a public-but-self-hosted location so
the Studio app can consume as dependencies. We expect this to continue
until we are again publishing releases to the NPM repo.

## Implementation details

Add a workflow for deploying self-hosted packages.

## Testing Instructions (or ideally a Blueprint)

We cannot test by manually running a new workflow until a version of the
workflow exists on the main branch. So we will have to commit this PR
without testing the workflow and follow up with a PR to fix the
workflow.

We can the next PR like this:
- Manually run the new workflow
- SSH into our package server and confirm the packages are deployed as
expected
- Create a local node package
- Add a dependency on one of the self-hosted packages that depends on
another (the `@wp-playground/cli` package URL should be good)
- `npm install`
- Confirm installation worked by testing `npx cli server`