GitHub: Cache node_modules per arch to avoid nx error (#70)
## Motivation for the change, related issues

We are seeing 'local cache artifact... was not been generated on this
machine' errors in GitHub Actions. It's not clear why we are seeing
these errors, but let's make sure we are not using a node_modules cache
generated on another architecture.

Related to #69

## Implementation details

This PR makes `runner.arch` a component of the cached node_modules key
in the prepare-playground action.

## Testing Instructions (or ideally a Blueprint)

Let's rely on the CI workflow. If that completes without issue, let's
merge this and then retry the website deployment workflow. Regardless of
whether this change fixes the issue, I believe cached node_modules
should only be used when their original runner.arch matches the current
runner.arch because `npm install` can install different things depending
on CPU architecture.