WASM: Support 64bit integers, drop PHP 7.0 and 7.1 (#74)
Enables 64bit arithmetic in all PHP.wasm builds.

In particular, this PR:

* Builds PHP with ` -D__x86_64__ ` to enable 64bit integers and links
the entire project with `-s WASM_BIGINT=1` to use the native JavaScript
BigInt for handling large integers in the JavaScript module.
* Adds the same ` -D__x86_64__ ` CLI arg to nearly all the WASM
dependencies. The only exceptions are `libwebp` and `libopenssl` where
adding that flag triggered compilation errors. We can very likely fix
them, but I didn't find a good reason to so we'll keep using 32bit ints
in two libraries until a good reason emerges.
* Drops support for PHP 7.0 and 7.1.

## Implementation

As it turns out, Emscripten and WebAssembly just support 64bit integers.
You can build this program and you'll see

```c
#include <stdio.h>
#include <stdlib.h>

int main() {
    printf("Data type sizes in this environment:\n");
    printf("Size of int: %d bytes\n", SIZEOF_INT);
    printf("Size of long: %d bytes\n", SIZEOF_LONG);
    printf("Size of long long: %d bytes\n", SIZEOF_LONG_LONG);
    printf("Size of int64_t: %zu bytes\n", sizeof(int64_t));
    return 0;
}
/**
 * Prints:
 * Size of int: 4 bytes
 * Size of long: 4 bytes
 * Size of long long: 8 bytes
 * Size of int64_t: 8 bytes
 */
```

I'm not sure when that happened. Or was it always like that? In any
case, PHP will use 64bit integers for its arithmetics and data types
when the `__x86_64__` preprocessor flag is present. This PR adds that
flag and rebuilds everything.

Unfortunately, when PHP detects `__x86_64__`, it also assumes it can
inline assembly instructions via `__asm__`. I don't believe this is
supported in Emscripten – we could make sure with the Google team. In
either case, I had to hack my way around it by setting the GLIBC macro
to 0 in specific files via different techniques.

We'll likely have to revisit and adjust this code in some way on every
major PHP release. In an ideal scenario, we would contribute a PR to PHP
core with preprocessor branches dedicated to WebAssembly builds.

### What's the net effect of this PR?

* PHP uses the `long long` data type for 64bit integers
* The JavaScript module uses BigInt class to handle large integers
* We **do not** use the memory64 spec, 64bit pointers, or 64bit address
space.

I can see two reasons to switch to the 64bit memory space:

* [Use more than 4GB of memory in the WASM
module](https://spidermonkey.dev/blog/2025/01/15/is-memory64-actually-worth-using.html).
* Potentially get a munmap()–supporting allocator and remove our custom
PHP memory allocator. We may or may not get it by building with Memory64
and I'd say it's out of scope of this PR.

Memory64 also has a downside: it [slows down the WASM program by about
10%](https://spidermonkey.dev/blog/2025/01/15/is-memory64-actually-worth-using.html):

> By reserving 4GB of memory for all 32-bit WebAssembly modules, it is
impossible to go out of bounds. The largest possible pointer value,
2^32-1, will simply land inside the reserved region of memory and trap.
This means that, when running 32-bit wasm on a 64-bit system, we can
omit all bounds checks
entirely[5](https://spidermonkey.dev/blog/2025/01/15/is-memory64-actually-worth-using.html#fn:5).
> 
> This optimization is impossible for Memory64. The size of the
WebAssembly address space is the same as the size of the host address
space. Therefore, we must pay the cost of bounds checks on every access,
and as a result, Memory64 is slower.

## Testing instructions

* Confirm all the CI tests pass. I've added a bunch of 64bit-specific
tests.