[PHP-wasm] Use statfs from NODEFS in the Node version (#94)
## Motivation for the change, related issues

This PR adds `statfs` support to the Emscripten root file system to
enable PHP-wasm Node to get real filesystem stats for the `/` path.

Currently, in PHP-wasm Node, PHP functions like `disk_total_space('/')`
return the default hardcoded value for MEMFS instead of the actual disk
space.

This happens because Emscripten automatically detects the filesystem for
a given path, and because the root path always uses the MEMFS
filesystem, Emscripten will use the MEMFS `statfs` implementation.

PHP-wasm Node includes two Emscripten filesystems, MEMFS and NODEFS.
MEMFS doesn't have access to the operating system's filesystem, so it
can only return a hardcoded value.
NODEFS has access to the OS filesystem through the `fs` module, and
specifically for `statfs` it uses `fs.statfsSync` to get real data.

## Implementation details

This PR uses Emscripten's `onRuntimeInitialized` event to override the
`statfs` implementation in the Emscripten root FS with the NODEFS
implementation of `statfs`.

It also defines the `FS.root.mount.opts.root` path as `.` to ensure the
root node has a path value defined.
Otherwise when we call `FS.statfs('/')` it would use the root node's
root path which is `undefined` and this would throw an error because it
would call `fs.statfsSync(undefined)` in the NODEFS implementation of
`statfs`.

I explored fixing this [issue in
Emscripten](https://github.com/emscripten-core/emscripten/pull/23912),
but I wasn't able to find a working solution there, so I ended up
patching it in Playground.

## Testing Instructions (or ideally a Blueprint)

- CI