Fix WP refresh workflows (#2190)
## Motivation for the change, related issues

The WP build refresh workflows are currently failing. Let's fix that.

## Implementation details

It looks like the following is happening:
- the build output is owned by root (perhaps due to Docker placing the
output in /root)
- the GitHub workflow is run by the user `runner`
- There is a permissions change in `wp-content/database` creation
[here](https://github.com/Automattic/sqlite-database-integration/blame/767d246522691c62f3e8ee316092e0cd781cede2/wp-includes/sqlite/class-wp-sqlite-db.php#L563)
where only the owner has permissions
- Git could not access all the new changes and failed

This PR just introduces `sudo chown` to fix that. There doesn't appear
to be an actual issue with the build script because there is no
permissions conflict when creating WordPress builds locally.

## Testing Instructions (or ideally a Blueprint)

- CI