[ Intl ] disable Intl functions in bootWordpress by default (#2247)
## Motivation for the change, related issues

The presence of the `Intl` extension is detected by checking if
functions exist in web mode, even when `withICU` is set to false or not
set at all.

Since `withICU` is set to false by default, the related functions should
be disabled.

This is a first and temporary attempt, as `intl` is currently static.
The best approach would be to generate a dynamic `intl` extension and
load it at runtime, if necessary.

## Implementation details

Alongside `withNetworking` or `networking`, add `withICU` and `icu` in
types and function parameters. Set `withICU` to false by default and
load a comprehensive list of `intl` functions to disable if `withICU` is
set to false.

To improve readability and maintainability, the list of disabled
functions is separated into a dedicated file.

## Testing Instructions (or ideally a Blueprint)

Tests are provided in
`packages/playground/website/playwright/e2e/blueprints.spec.ts`

---------

Co-authored-by: Bero <berislav.grgicak@gmail.com>