Allow the WP update and changelog workflows to commit to trunk again (#2377)
## Motivation for the change, related issues

The WP build and changelog update workflows are currently blocked from
committing to trunk because it is a protected branch.

## Implementation details
This PR fixes that by restoring use of GH_ACTOR and GH_TOKEN in those
workflows.

I previously removed use of the occasionally-expiring Personal Access
Token (PAT), not understanding that it was required for pushing to
protected branches. At the time we were not protecting trunk in the that
repo, and using the default token provided in the GH workflow context
worked just fine.

Once we switched back to using branch protection, these workflows
started breaking.

For a GitHub Actions workflow to push to a protected branch we need to
do one of the following:
- Use a PAT and update it before it expires
- Use a Deploy Key (doesn't expire but has a lot of power if it falls
into the wrong hands)
- Create a GitHub app at the organization level, add it to the repo,
create an app token during workflow execution, and use that. This is a
bit annoying since we don't (or at least I don't) have access to create
an app within the WordPress GH org, so we'd have to request the app and
then jump through hoops to use it in each workflow.
- Create a PR within the workflow and use the GH REST API (even via GH
CLI) to merge it

Right now, we have a PAT that is working for the
sqlite-database-integration update, so I'm just restoring that for the
other workflows.

## Testing Instructions (or ideally a Blueprint)

Merge and run workflows on trunk.