[Node] Gracefully handle connection errors in the outbound network proxy (#2370)
## Motivation for the change, related issues

As discovered in https://github.com/Automattic/studio/pull/1491, adding
a timeout before closing the network connection in the outbound network
proxy prevents freezing the PHP runtime when the target cannot be
reached.

This PR adds such a timeout.

The exact reason behind this behavior isn't well understood yet. The
freeze is not deterministic or easy to reproduce, which is why this PR
does not ship any tests. We should keep an eye on this and, if it ever
comes up again, invest some time in a proper deep dive to understand all
the factors at play.

Related to https://github.com/Automattic/studio/pull/1491

cc @gavande1 @mho22