Adding the docs meta description to the main folder (#2504)
## Motivation for the change, related issues
This pull request includes a meta description for the pages present in
the `main` folder. The goal of this change is to include a meaningful
description about the content present on the page to help search engines
and improve the click rate on content shared on social media.

This pull request also includes the meta description for Brazilian
Portuguese pages

This task is related to issue #2500 

Adding the docs meta description to the main folder