[WordPress][CLI] Allow requesting internal URLs via http_request_host_is_external (#2569)
## Motivation for the change, related issues

WordPress uses gethostbyname() inside wp_http_validate_url() to decide
if a host is external.
In Emscripten, this resolves to a private network IP, which causes
WordPress to reject all outgoing requests.

This change adds the http_request_host_is_external filter returning true
to bypass the check. While unsafe in production, it’s acceptable in
Playground because the environment already allows arbitrary code
execution (e.g. Blueprints, CLI).

Also added a test to confirm the filter is applied after WordPress
boots.

## Future work

Ideally, `gethostbyname()` would just work. While it's possible to shim
it in Node.js, the only solution in the browser seems to be issuing
`fetch()` to a PHP script to would perform the DNS resolution. That
would be super slow.

Realistically, we can either:

* Implement gethostbyname in Node.js and not implement it in the web
version of Playground
* Apply the same filter in Node.js as we already do on the web – which
is at least consistent

This PR does the latter.

Related to #400 

## Testing Instructions (or ideally a Blueprint)

CI