Playground CLI: Mount /wordpress, /internal, and /tmp dirs within real temporary dir (#2446)
## Motivation for the change, related issues

There are two reasons to mount a real directory as `/internal/shared`:
1. We intend `/internal/shared` to be truly shared between all PHP
instances. Issue:
  #2301.
3. Zipping the primary worker `/internal` dir and unzipping for
secondary workers is making multi-worker CLI startup quite slow. Issue:
  #2419.

Once we looked at doing the above, we saw that it would be good for
workers to share the same real FS by default. It made it easier for
users to start a multi-worker Playground server, and those workers would
all naturally share the same underlying FS state.

## Implementation details

- When launching, Playground CLI:
- Creates a native temp dir that is configured to be cleaned up before
the application exits.
- Launches a non-blocking cleanup of any unused Playground temp dirs
that are:
    - Older than two days
    - No longer associated with a running process
- We create the following subdirectories in the temp dir:
  - `/internal`
  - `/wordpress`
- `/home` - We don't use this AFAIK, but since it is assumed to exist in
the emscripten-generated php-wasm JS, it seemed good to make sure it is
also backed by a native dir.
  - `/tmp`
- The real path for the `/internal` dir is passed to php-wasm init where
it handles mounting and creation of internal subdirs.
- If the user does not provide mounts for `/wordpress`, `/home`, or
`/tmp`, we mount the temp subdirs in their place.

## Testing Instructions (or ideally a Blueprint)

- CI

---------

Co-authored-by: Adam Zieliński <adam@adamziel.com>