[Website] Isolate resolveBlueprintFromURL() calls (#2654)
## Motivation for the change, related issues

Moves all the `resolveBlueprintFromURL()` into a single state function
to set the stage for Blueprints v2. Version 2 will handle merging Query
API overrides with the Blueprint using PHP logic, not TypeScript logic.
This PR is the first step there – it moves all the TypeScript parsing
and merging logic into a single place to make the upcoming refactoring
easier.

A part of #2586. 

## Testing Instructions (or ideally a Blueprint)

CI