[Blueprints] Use _SERVER['HTTPS'] in the import step (#2802)
## Motivation for the change, related issues

Importing WXR content with URL rewriting enabled replaces the protocol
in some URLs from `https` to `http`.

## Implementation details

Sets `$_SERVER['HTTPS'] = "on"` in the `importWxr` Blueprint step. The
`php.run()` method called in the `importWxr` step does not set it
automatically – that's a role of the request handler. The WXR importer
relies on the `get_site_url()` WordPress function, which weirdly takes
the `site_url` option and replaces the protocol based on the
`$_SERVER['https']` value.

## Testing Instructions (or ideally a Blueprint)

Import this WXR file via `?importWxr` query parameter and confirm all
the URLs are https:


https://gist.github.com/adamziel/a3d1a1941608e068e4b3d557ee2e8202/raw/aa659a130e1104cedeea8a63155a117dcb693dc8/tt25export.xml

The tricky part is, you need to serve Playground on a https url for this
to work.

cc @bph