[docs] Removing GitHub proxy references (#3009)
This pull request updates the documentation for plugin and theme
developers to reflect the deprecation of the GitHub proxy in favor of
the new `git:directory` resource for loading plugins and themes from
GitHub repositories in WordPress Playground. It also improves clarity,
updates example blueprints, and enhances the Spanish translation for
plugin developer documentation.