[Website] Skip WordPress installation when WordPress files already exist (#3065)
## Summary

Without this PR https://playground.wordpress.net/ always runs the
WordPress installer even when running a mini fight site that's already
pre-installed and shipped to the SQLite database. This adds around 600
milliseconds to every playground boot on a beefy MacBook Pro.

With this PR, that WordPress installer request is skipped when possible.
How? It sets the `wordpressInstallMode` to
`install-from-existing-files-if-needed` that makes the installer
optional.

## Test plan

- Load a saved Playground site and verify it still boots correctly
- Create a new Playground and verify WordPress installs as expected
- Observe faster boot times when loading existing sites